+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.1
 
 * fixed selecting largest icon
+* fixed fallback to /favicon.ico if index cannot be loaded
 
 ## 1.0.0